Added exploratory data analysis (EDA) methods.
In the ExploreFeatures class in data.py, I added
some basic exploratory data analysis methods. Each
of these generates a plot showing the relationship
between the numeric and categoric feeatures and
the target values.

This is intended to be used as an initial first
look into a dataset to see if there are any
obvious patterns or issues.